+++ resources/views/emails/ticket-no-tickets.blade.php
@@ -1,7 +1,7 @@
 @component('mail::message', ['email' => $email])
 <p>Hey there!</p>
-<p>We couldn’t find any tickets linked to this email address.</p>
-<p>If this seems incorrect, check you entered the same email used for the ticket recipient details.</p>
+<p>We weren’t able to find any tickets linked to this email address.</p>
+<p>It’s possible the tickets were booked using a different email. If you have another address you may have used, feel free to try that. If something doesn’t look right, just reply to this email and we’ll help sort it out.</p>
 
 <p>Warm regards,</p>
 <p>—James 😁</p>